migrate sidebar types
